%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  fitNSA.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{fitNSA.matrix}
\alias{fitNSA.matrix}
\alias{fitNSA.matrix}

 \alias{fitNSA}

 \title{Find normal regions}

 \description{
  Find normal regions.
 }

 \usage{\method{fitNSA}{matrix}(data, ...)}

 \arguments{
  \item{data}{An JxI \code{\link[base]{numeric}} \code{\link[base]{array}}, where J is the number of SNPs,
          and I is the number of samples.}
  \item{...}{Additional arguments passed to internal
             \code{calmate().}}
 }

 \value{
   Returns an JxI \code{\link[base]{numeric}} \code{\link[base]{array}}.
 }                                 
\keyword{methods}
