%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{truncateThetaAB}
\alias{truncateThetaAB.array}
\alias{allocateOutputDataSets}
\alias{allocateOutputDataSets.NSANormalization}
\alias{allocateOutputDataSets.SNPsNormalization}
\alias{allocateOutputDataSets.SampleNormalization}
\alias{as.character.NSANormalization}
\alias{as.character.SNPsNormalization}
\alias{as.character.SampleNormalization}
\alias{getAsteriskTags}
\alias{getAsteriskTags.NSANormalization}
\alias{getAsteriskTags.SNPsNormalization}
\alias{getAsteriskTags.SampleNormalization}
\alias{getDataSets}
\alias{getDataSets.NSANormalization}
\alias{getDataSets.SNPsNormalization}
\alias{getDataSets.SampleNormalization}
\alias{getFullName}
\alias{getFullName.NSANormalization}
\alias{getFullName.SNPsNormalization}
\alias{getFullName.SampleNormalization}
\alias{getInputDataSet}
\alias{getInputDataSet.NSANormalization}
\alias{getInputDataSet.SNPsNormalization}
\alias{getInputDataSet.SampleNormalization}
\alias{getName.NSANormalization}
\alias{getName.SNPsNormalization}
\alias{getName.SampleNormalization}
\alias{getOutputDataSet}
\alias{getOutputDataSet.NSANormalization}
\alias{getOutputDataSet.SNPsNormalization}
\alias{getOutputDataSet.SampleNormalization}
\alias{getPath.NSANormalization}
\alias{getPath.SampleNormalization}
\alias{getPath.SNPsNormalization}
\alias{getRootPath}
\alias{getRootPath.NSANormalization}
\alias{getRootPath.SampleNormalization}
\alias{getRootPath.SNPsNormalization}
\alias{getTags}
\alias{getTags.NSANormalization}
\alias{getTags.SNPsNormalization}
\alias{getTags.SampleNormalization}
\alias{nbrOfFiles}
\alias{nbrOfFiles.NSANormalization}
\alias{nbrOfFiles.SNPsNormalization}
\alias{nbrOfFiles.SampleNormalization}
\alias{process}
\alias{process.SNPsNormalization}
\alias{setTags}
\alias{setTags.NSANormalization}
\alias{setTags.SNPsNormalization}
\alias{setTags.SampleNormalization}
\alias{findArraysTodo}
\alias{findArraysTodo.NSANormalization}
\alias{findArraysTodo.SampleNormalization}
\alias{findUnitsTodo}
\alias{findUnitsTodo.NSANormalization}
\alias{findUnitsTodo.SNPsNormalization}
\alias{findUnitsTodo.SampleNormalization}
\alias{getOutputDataSets}
\alias{getOutputDataSets.NSANormalization}
\alias{getOutputDataSets.SNPsNormalization}
\alias{getOutputDataSets.SampleNormalization}
\alias{truncateFracB}
\alias{truncateFracB.array}
\alias{truncateFracB.matrix}
\alias{truncateThetaAB.matrix}
\alias{fitSNPsN}
\alias{fitSample}
\alias{fitOneNSA}
\alias{fitNSAcnPs}
\alias{fitNSAcnPs.matrix}
\alias{LHMatrix}
\alias{getReferences}
\alias{getReferences.NSANormalization}
\alias{getReferences.SampleNormalization}
\alias{getReferences.SNPsNormalization}
\alias{cat}
\alias{sapply}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}


\keyword{documentation}
\keyword{internal}
