%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{NSA-package}
\alias{NSA-package}
\alias{NSA}
\docType{package}

\title{Package NSA}


 \description{
   Post-normalization of total copy-number estimates.

   This package should be considered to be in an alpha phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires...
 }

 \section{Installation and updates}{
   TBA.
 }

 \section{To get started}{
  \enumerate{
   \item To process SNP and non-polymorphic signals, see \code{\link{NSAByTotalAndFracB}}.
   \item For processing data in the aroma framework, see \code{\link{NSANormalization}}.
  }
 }

 \author{
   Maria Ortiz-Estevez <mortiz@ceit.es>,
   Ander Aramburu <aaramburu@ceit.es>,
   Pierre Neuvial <pierre@stat.berkeley.edu>,
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.
 }

 \references{
   TBA.
 }
\keyword{package}
